requirements.txt
aiogram>=3.23.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
pymupdf>=1.26.0
pillow>=10.2.0
pypdf>=4.0.0